The Banking ML application to find possible clients using ML
This Project contains simple ML to find the most suitable clients who are willing to take the offer based on the gathered input. The Source is freely and openly available and the code is self explanatory. The data set is 10% of the original data set and only contains a portion of the code. but that is enough to run the code.


**RUN REQUIREMENTS**
-- Python 3.5 and above
-- Jupiter Notebook